REFACTOR: Split documents source files
